requirements.txt
python == 3.7.6
pytorch == 1.6.0
numpy >= 1.15.4